[ALPS03885364] WLAN: fix coverity Printf arg type mismatch
[Detail]
fix coverity Scanned Code Defect, Printf arg type mismatch

MTK-Commit-Id: b680df10d256d05b84186950badb762b0bcbd4fd

CR-Id: ALPS03885364
Feature: WIFI
(cherry picked from commit 838ec782b1cf5bf5a870d194f1971e1da689d6b0)

+++ os/linux/hif/common/hal_pdma.c
@@ -1042,7 +1042,8 @@ void halRxReceiveRFBs(IN struct ADAPTER *prAdapter, uint32_t u4Port)
 			HAL_RX_STATUS_GET_PKT_TYPE(prRxStatus);
 		DBGLOG(RX, TRACE, "ucPacketType = %u, ucSecMode = %u\n",
 				  prSwRfb->ucPacketType,
-				  HAL_RX_STATUS_GET_SEC_MODE(prRxStatus));
+				  (uint8_t)HAL_RX_STATUS_GET_SEC_MODE(
+					prRxStatus));
 
 		if (prSwRfb->ucPacketType == RX_PKT_TYPE_MSDU_REPORT) {
 			nicRxProcessMsduReport(prAdapter, prSwRfb);
